修复new api获取渠道报错的问题

+++ src/main/java/github/gpt/api/sync/mapper/ChannelMapper.java
@@ -27,7 +27,7 @@ public class ChannelMapper {
 
         channel.setStatus(1); // Enabled by default
         channel.setModels(group.getTestModel()); // Use test_model as the default models
-        channel.setPriority((long) group.getId());
+        channel.setPriority(group.getId());
 
         return channel;
     }
